Updating Information about the user: - Backend: Adding a new DTO so that we can set a new username, new profile picture. Adding new services and Interfaces: to update informations. Adding new controllers: Three Put (Username, ProfilePicture and Status) and one Get (ProfilePicture).
- Frontend:
Adding a new component settings, where you can change information about yourself.
Adding the services and Interfaces that link with the backend.
Changing a bit of the home component: Adding a logout feature, and a router with navigable links (To implement later).
Adding much more validators for the reset and forgot password, that were forgotten in the previous commit.